Change price type from int to long across variant-related classes for better precision.

+++ backend/services/product-service/src/main/java/vn/edu/uit/msshop/product/variant/adapter/in/web/request/UpdateVariantInfoRequest.java
@@ -5,7 +5,7 @@ import java.util.List;
 import vn.edu.uit.msshop.product.shared.adapter.in.web.request.ChangeRequest;
 
 public record UpdateVariantInfoRequest(
-        ChangeRequest<Integer> price,
+        ChangeRequest<Long> price,
         ChangeRequest<List<String>> traits,
         long version) {
 }
